MSBG-73 linter installed
