+++ .eslintrc
@@ -1,25 +1,109 @@
 {
-  "extends": "eslint:recommended",
-  "parser": "babel-eslint",
-  "env": {
-    "browser": true,
-    "node": true
-  },
-  "plugins": [
-    "react"
-  ],
-  "rules": {
-    "no-debugger": 0,
-    "no-console": 0,
-    "new-cap": 0,
-    "strict": 0,
-    "no-underscore-dangle": 0,
-    "no-use-before-define": 0,
-    "eol-last": 0,
-    "quotes": [2, "single"],
-    "jsx-quotes": [1, "prefer-single"],
-    "react/jsx-no-undef": 1,
-    "react/jsx-uses-react": 1,
-    "react/jsx-uses-vars": 1
-  }
+    "extends": "airbnb",
+    "parser": "babel-eslint",
+    "env": {
+        "browser": true,
+        "node": true
+    },
+    "plugins": [
+        "react"
+    ],
+    "rules": {
+        "indent": [
+            "error",
+            4,
+            {
+                "SwitchCase": 1
+            }
+        ],
+        "import/prefer-default-export": "off",
+        "import/no-extraneous-dependencies": [
+            "warn",
+            {
+                "devDependencies": [
+                    "spec/**",
+                    "test/**",
+                    "tests/**",
+                    "**/__tests__/**"
+                ],
+                "optionalDependencies": false
+            }
+        ],
+        "jsx-a11y/label-has-for": "off",
+        "jsx-a11y/no-static-element-interactions": "off",
+        "object-curly-spacing": [
+            "error",
+            "never"
+        ],
+        "comma-dangle": [
+            "error",
+            "never"
+        ],
+        "class-methods-use-this": [
+            "warn",
+            {
+                "exceptMethods": [
+                    "render"
+                ]
+            }
+        ],
+        "max-len": [
+            "error",
+            120,
+            2,
+            {
+                "ignoreUrls": true,
+                "ignoreComments": false,
+                "ignoreRegExpLiterals": true,
+                "ignoreStrings": true,
+                "ignoreTemplateLiterals": true
+            }
+        ],
+        "no-use-before-define": [
+            "error",
+            {
+                "functions": false
+            }
+        ],
+        "no-underscore-dangle": [
+            "off"
+        ],
+        "react/jsx-indent": [
+            "error",
+            4
+        ],
+        "react/jsx-indent-props": [
+            "error",
+            4
+        ],
+        "no-duplicate-imports": "off",
+        // we use eslint-import-plugin instead
+        "react/forbid-prop-types": [
+            "error",
+            {
+                "forbid": [
+                    "any",
+                    "array"
+                ]
+            }
+        ],
+        "react/prefer-stateless-function": [
+            "off",
+            {
+                "ignorePureComponents": true
+            }
+        ],
+        "react/prop-types": [
+            "warn",
+            {
+                "ignore": [
+                    "className",
+                    "children",
+                    "location",
+                    "props"
+                ],
+                "customValidators": []
+            }
+        ]
+    }
 }
