+++ package.json
@@ -11,16 +11,18 @@
   "devDependencies": {
     "autoprefixer": "^6.3.1",
     "babel-core": "^6.4.5",
-    "babel-eslint": "^4.1.6",
+    "babel-eslint": "^7.2.3",
     "babel-loader": "^6.2.1",
     "babel-plugin-transform-runtime": "^6.4.3",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-0": "^6.3.13",
     "css-loader": "^0.23.1",
-    "eslint": "^1.10.3",
-    "eslint-loader": "^1.2.1",
-    "eslint-plugin-react": "^3.16.1",
+    "eslint": "^4.3.0",
+    "eslint-loader": "^1.9.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
+    "eslint-plugin-react": "^7.1.0",
     "express": "^4.13.4",
     "npm-install-webpack-plugin": "^2.0.2",
     "postcss-loader": "^0.8.0",
@@ -35,6 +37,7 @@
   "dependencies": {
     "babel-polyfill": "^6.3.14",
     "babel-runtime": "^6.3.19",
+    "eslint-config-airbnb": "^15.1.0",
     "lodash": "^4.17.4",
     "react": "^0.14.6",
     "react-dom": "^0.14.6",
